+++ src/entities/clothes/index.ts
@@ -1,8 +1,4 @@
-export {
-  deleteLookbookItem,
-  getAllLookbookListByWeather,
-  saveLookbook,
-} from './api/lookbook';
+export { getAllLookbookListByWeather } from './api/lookbook';
 
 export type { ClothesIconNames, ClothesType } from './model/types';
